examples/news/requirements.txt
../../
flask
newspaper3k
sentence-transformers
rapidfuzz